produxman-llm/requirements.txt
flask
flask-cors
waitress
flask-basicauth
requests
chromadb
pypdf
python-dotenv
